Move resolvers folder upper
